requirements.txt
asgiref==3.3.1
Django==3.1.7
gunicorn==20.1.0
Pillow==8.2.0
pypng==0.0.20
PyQRCode==1.2.1
pytz==2021.1
sqlparse==0.4.1
whitenoise==5.2.0
